Refactor deep sleep handling in IoT device code (#16)
The deep sleep disable mechanism in our IoT device services is replaced with a wakeup mechanism. The GPIO pin connected to waking up the device from deep sleep is now configured through the DeepSleepDataProviderConfiguration. Use of this pin is adjusted throughout the codebase, also simplifying some constructors and error handling. Lastly, the logic is adjusted for determining if the device should go into deep sleep based on pin state.